+++ packages/dash-table/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+[#488](https://github.com/plotly/dash-table/pull/488)
+- Update table build for use as a library and make consistent with other Dash repos
+
 ## [4.0.0] - 2019-06-20
 ### Changed
 [#446](https://github.com/plotly/dash-table/pull/446)